maxboom_backend/requirements.txt
asgiref==3.7.2
Django==3.2.3
djangorestframework==3.12.4
Pillow==10.0.0
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
pytils==0.4.1
slugify==0.0.1
sorl-thumbnail==12.9.0
django-filter==23.3
djoser==2.2.0
drf-spectacular==0.26.5
django-cors-headers==4.2.0
deprecated==1.2.14
distro==1.8.0
netaddr==0.9.0
wrapt==1.15.0
yookassa==2.4.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.24
